+++ jackssh-presentation/Sources/Presentation/Connection/OpenClawDashboard/Model/OpenClawDashboardEffect.swift
@@ -1,0 +1,4 @@
+public enum OpenClawDashboardEffect: Equatable {
+    case none
+    case showError(String)
+}